Integrated Stripe Payment API
1. Integrated with stripe payment API
2. Adding donor delivery feature

+++ src/main/java/com/utar/uhauction/model/dto/BidDTO.java
@@ -3,6 +3,7 @@ package com.utar.uhauction.model.dto;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 public class BidDTO implements Serializable {
@@ -12,5 +13,5 @@ public class BidDTO implements Serializable {
     private String item_id;
 
 
-    private Double amount;
+    private Long amount;
 }
